began bls job industry refactor to table format
